Move counterpart funding to its own file
