+++ maediprojects/__init__.py
@@ -69,6 +69,7 @@ def register_blueprints(app):
     app.register_blueprint(views.exports.blueprint)
     app.register_blueprint(views.reports.blueprint)
     app.register_blueprint(views.management.blueprint)
+    app.register_blueprint(views.counterpart_funding.blueprint)
 
 
 def register_errorhandlers(app):
