+++ maediprojects/views/__init__.py
@@ -1,1 +1,3 @@
-from maediprojects.views import activities, api, iati, codelists, users, exports, reports, management
+from maediprojects.views import activities, api, iati, \
+	codelists, users, exports, reports, \
+	management, counterpart_funding
